chore(1-3441): new strategy draggable item
