+++ frontend/src/component/feature/FeatureView/FeatureOverview/FeatureOverviewEnvironments/FeatureOverviewEnvironment/EnvironmentAccordionBody/StrategyDraggableItem/StrategyDraggableItem.tsx
@@ -4,92 +4,13 @@ import {
     useRef,
     type ReactNode,
 } from 'react';
-import { Box, useMediaQuery, useTheme } from '@mui/material';
-import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
-import type { IFeatureEnvironment } from 'interfaces/featureToggle';
+import { Box } from '@mui/material';
 import type { IFeatureStrategy } from 'interfaces/strategy';
-import { useRequiredPathParam } from 'hooks/useRequiredPathParam';
-import {
-    useStrategyChangesFromRequest,
-    type UseStrategyChangeFromRequestResult,
-} from './StrategyItem/useStrategyChangesFromRequest';
-import { ChangesScheduledBadge } from 'component/changeRequest/ModifiedInChangeRequestStatusBadge/ChangesScheduledBadge';
-import type { IFeatureChange } from 'component/changeRequest/changeRequest.types';
-import { Badge } from 'component/common/Badge/Badge';
-import {
-    type ScheduledChangeRequestViewModel,
-    useScheduledChangeRequestsWithStrategy,
-} from 'hooks/api/getters/useScheduledChangeRequestsWithStrategy/useScheduledChangeRequestsWithStrategy';
-import {
-    StrategyItem,
-    StrategyItemNoProject,
-} from './StrategyItem/StrategyItem';
-
-interface IStrategyDraggableItemProps {
-    strategy: IFeatureStrategy;
-    environmentName: string;
-    index: number;
-    otherEnvironments?: IFeatureEnvironment['name'][];
-    isDragging?: boolean;
-    onDragStartRef?: (
-        ref: RefObject<HTMLDivElement>,
-        index: number,
-    ) => DragEventHandler<HTMLButtonElement>;
-    onDragOver?: (
-        ref: RefObject<HTMLDivElement>,
-        index: number,
-    ) => DragEventHandler<HTMLDivElement>;
-    onDragEnd?: () => void;
-}
+import { StrategyItem } from './StrategyItem/StrategyItem';
 
 const onDragNoOp = () => () => {};
 
-export const StrategyDraggableItem = ({
-    strategy,
-    index,
-    environmentName,
-    otherEnvironments,
-    isDragging,
-    onDragStartRef = onDragNoOp,
-    onDragOver = onDragNoOp,
-    onDragEnd = onDragNoOp,
-}: IStrategyDraggableItemProps) => {
-    const projectId = useRequiredPathParam('projectId');
-    const featureId = useRequiredPathParam('featureId');
-    const ref = useRef<HTMLDivElement>(null);
-    const strategyChangesFromRequest = useStrategyChangesFromRequest(
-        projectId,
-        featureId,
-        environmentName,
-        strategy.id,
-    );
-
-    const { changeRequests: scheduledChangesUsingStrategy } =
-        useScheduledChangeRequestsWithStrategy(projectId, strategy.id);
-
-    return (
-        <Box
-            key={strategy.id}
-            ref={ref}
-            onDragOver={onDragOver(ref, index)}
-            sx={{ opacity: isDragging ? '0.5' : '1' }}
-        >
-            <StrategyItem
-                strategy={strategy}
-                environmentId={environmentName}
-                otherEnvironments={otherEnvironments}
-                onDragStart={onDragStartRef(ref, index)}
-                onDragEnd={onDragEnd}
-                headerChildren={renderHeaderChildren(
-                    strategyChangesFromRequest,
-                    scheduledChangesUsingStrategy,
-                )}
-            />
-        </Box>
-    );
-};
-
-type Props = {
+type StrategyDraggableItemProps = {
     actions: ReactNode;
     strategy: IFeatureStrategy;
     index: number;
@@ -105,7 +26,7 @@ type Props = {
     onDragEnd?: () => void;
 };
 
-export const StrategyDraggableItemNoEnv = ({
+export const StrategyDraggableItem = ({
     strategy,
     index,
     isDragging,
@@ -113,7 +34,7 @@ export const StrategyDraggableItemNoEnv = ({
     onDragOver = onDragNoOp,
     onDragEnd = onDragNoOp,
     actions,
-}: Props) => {
+}: StrategyDraggableItemProps) => {
     const ref = useRef<HTMLDivElement>(null);
 
     return (
@@ -123,7 +44,7 @@ export const StrategyDraggableItemNoEnv = ({
             onDragOver={onDragOver(ref, index)}
             sx={{ opacity: isDragging ? '0.5' : '1' }}
         >
-            <StrategyItemNoProject
+            <StrategyItem
                 actions={actions}
                 strategy={strategy}
                 onDragStart={onDragStartRef(ref, index)}
@@ -131,66 +52,4 @@ export const StrategyDraggableItemNoEnv = ({
             />
         </Box>
     );
-};
-
-const ChangeRequestStatusBadge = ({
-    change,
-}: {
-    change: IFeatureChange | undefined;
-}) => {
-    const theme = useTheme();
-    const isSmallScreen = useMediaQuery(theme.breakpoints.down('sm'));
-
-    if (isSmallScreen) {
-        return null;
-    }
-
-    return (
-        <Box sx={{ mr: 1.5 }}>
-            <ConditionallyRender
-                condition={change?.action === 'updateStrategy'}
-                show={<Badge color='warning'>Modified in draft</Badge>}
-            />
-            <ConditionallyRender
-                condition={change?.action === 'deleteStrategy'}
-                show={<Badge color='error'>Deleted in draft</Badge>}
-            />
-        </Box>
-    );
-};
-
-const renderHeaderChildren = (
-    changes?: UseStrategyChangeFromRequestResult,
-    scheduledChanges?: ScheduledChangeRequestViewModel[],
-): JSX.Element[] => {
-    const badges: JSX.Element[] = [];
-    if (changes?.length === 0 && scheduledChanges?.length === 0) {
-        return [];
-    }
-
-    const draftChange = changes?.find(
-        ({ isScheduledChange }) => !isScheduledChange,
-    );
-
-    if (draftChange) {
-        badges.push(
-            <ChangeRequestStatusBadge
-                key={`draft-change#${draftChange.change.id}`}
-                change={draftChange.change}
-            />,
-        );
-    }
-
-    if (scheduledChanges && scheduledChanges.length > 0) {
-        badges.push(
-            <ChangesScheduledBadge
-                key='scheduled-changes'
-                scheduledChangeRequestIds={scheduledChanges.map(
-                    (scheduledChange) => scheduledChange.id,
-                )}
-            />,
-        );
-    }
-
-    return badges;
 };
