+++ frontend/src/component/feature/FeatureView/FeatureOverview/FeatureOverviewEnvironments/FeatureOverviewEnvironment/EnvironmentAccordionBody/StrategyDraggableItem/StrategyItem/StrategyItem.tsx
@@ -1,129 +1,18 @@
 import type { DragEventHandler, FC, ReactNode } from 'react';
-import Edit from '@mui/icons-material/Edit';
-import { Link } from 'react-router-dom';
-import type { IFeatureEnvironment } from 'interfaces/featureToggle';
 import type { IFeatureStrategy } from 'interfaces/strategy';
-import PermissionIconButton from 'component/common/PermissionIconButton/PermissionIconButton';
-import { UPDATE_FEATURE_STRATEGY } from 'component/providers/AccessProvider/permissions';
-import { formatEditStrategyPath } from 'component/feature/FeatureStrategy/FeatureStrategyEdit/FeatureStrategyEdit';
-import { useRequiredPathParam } from 'hooks/useRequiredPathParam';
 import { StrategyExecution } from './StrategyExecution/StrategyExecution';
-import { CopyStrategyIconMenu } from './CopyStrategyIconMenu/CopyStrategyIconMenu';
-import MenuStrategyRemove from './MenuStrategyRemove/MenuStrategyRemove';
 import SplitPreviewSlider from 'component/feature/StrategyTypes/SplitPreviewSlider/SplitPreviewSlider';
-import { Box, useMediaQuery, useTheme } from '@mui/material';
+import { Box } from '@mui/material';
 import { StrategyItemContainer as NewStrategyItemContainer } from 'component/common/StrategyItemContainer/StrategyItemContainer';
-import { useScheduledChangeRequestsWithStrategy } from 'hooks/api/getters/useScheduledChangeRequestsWithStrategy/useScheduledChangeRequestsWithStrategy';
-import { useStrategyChangesFromRequest } from './useStrategyChangesFromRequest';
-import { ChangesScheduledBadge } from 'component/changeRequest/ModifiedInChangeRequestStatusBadge/ChangesScheduledBadge';
-import { ChangeRequestDraftStatusBadge } from 'component/changeRequest/ChangeRequestStatusBadge/ChangeRequestDraftStatusBadge';
-interface IStrategyItemProps {
-    environmentId: string;
-    strategy: IFeatureStrategy;
-    onDragStart?: DragEventHandler<HTMLButtonElement>;
-    onDragEnd?: DragEventHandler<HTMLButtonElement>;
-    otherEnvironments?: IFeatureEnvironment['name'][];
-    headerChildren?: JSX.Element[] | JSX.Element;
-}
 
-export const StrategyItem: FC<IStrategyItemProps> = ({
-    environmentId,
-    strategy,
-    onDragStart,
-    onDragEnd,
-    otherEnvironments,
-    headerChildren,
-}) => {
-    const projectId = useRequiredPathParam('projectId');
-    const featureId = useRequiredPathParam('featureId');
-
-    const editStrategyPath = formatEditStrategyPath(
-        projectId,
-        featureId,
-        environmentId,
-        strategy.id,
-    );
-
-    const strategyChangesFromRequest = useStrategyChangesFromRequest(
-        projectId,
-        featureId,
-        environmentId,
-        strategy.id,
-    );
-
-    const { changeRequests: scheduledChanges } =
-        useScheduledChangeRequestsWithStrategy(projectId, strategy.id);
-
-    const draftChange = strategyChangesFromRequest?.find(
-        ({ isScheduledChange }) => !isScheduledChange,
-    );
-
-    const theme = useTheme();
-    const isSmallScreen = useMediaQuery(theme.breakpoints.down('sm'));
-
-    return (
-        <StrategyItemNoProject
-            strategy={strategy}
-            onDragStart={onDragStart}
-            onDragEnd={onDragEnd}
-            actions={
-                <>
-                    {draftChange && !isSmallScreen ? (
-                        <ChangeRequestDraftStatusBadge
-                            sx={{ mr: 1.5 }}
-                            changeAction={'updateStrategy'}
-                        />
-                    ) : null}
-
-                    {scheduledChanges &&
-                    scheduledChanges.length > 0 &&
-                    !isSmallScreen ? (
-                        <ChangesScheduledBadge
-                            scheduledChangeRequestIds={(
-                                scheduledChanges ?? []
-                            ).map((scheduledChange) => scheduledChange.id)}
-                        />
-                    ) : null}
-                    {otherEnvironments && otherEnvironments?.length > 0 ? (
-                        <CopyStrategyIconMenu
-                            environmentId={environmentId}
-                            environments={otherEnvironments as string[]}
-                            strategy={strategy}
-                        />
-                    ) : null}
-                    <PermissionIconButton
-                        permission={UPDATE_FEATURE_STRATEGY}
-                        environmentId={environmentId}
-                        projectId={projectId}
-                        component={Link}
-                        to={editStrategyPath}
-                        tooltipProps={{
-                            title: 'Edit strategy',
-                        }}
-                        data-testid={`STRATEGY_EDIT-${strategy.name}`}
-                    >
-                        <Edit />
-                    </PermissionIconButton>
-                    <MenuStrategyRemove
-                        projectId={projectId}
-                        featureId={featureId}
-                        environmentId={environmentId}
-                        strategy={strategy}
-                    />
-                </>
-            }
-        />
-    );
-};
-
-type Props = {
+type StrategyItemProps = {
     actions: ReactNode;
     strategy: IFeatureStrategy;
     onDragStart?: DragEventHandler<HTMLButtonElement>;
     onDragEnd?: DragEventHandler<HTMLButtonElement>;
 };
 
-export const StrategyItemNoProject: FC<Props> = ({
+export const StrategyItem: FC<StrategyItemProps> = ({
     strategy,
     onDragStart,
     onDragEnd,
